alix3d3.conf: Disable SERIAL and enable explicit configuration
At first ia32-base.conf was extended. However, this configuration is
generic for all intel architecture and it did not fit out needs.
ia32-base set SERIAL_CONSOLE which we do not want getty to bind. The
board has VGA that are used for console. The serial port will be used
for other purposes.

Move XSERVER option from base configuration as wel.